# backend/cmd/server/main.go
package main

import (
	"fmt"
	"log"
	"net/http"

	_ "crm-backend-go/docs" // Swagger docs generated by swag
	"crm-backend-go/internal/config"
	"crm-backend-go/internal/database"
	"crm-backend-go/internal/handlers"
	"crm-backend-go/internal/middleware"
	"crm-backend-go/internal/repository"

	"github.com/gin-gonic/gin"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
)

// @title Universal CRM - Lead Management API (Go)
// @version 2.0.0
// @description High-performance Go + Gin + PostgreSQL 17 CRM Lead Management backend.
// @BasePath /
// @securityDefinitions.apikey BearerAuth
// @in header
// @name Authorization
func main() {
	cfg := config.LoadConfig()

	db, err := database.ConnectDB(cfg.DatabaseURL)
	if err != nil {
		log.Fatalf("Database connection failure: %v", err)
	}
	defer db.Close()

	if err := db.Migrate(); err != nil {
		log.Fatalf("Migration failure: %v", err)
	}

	if err := db.Seed(); err != nil {
		log.Fatalf("Seeder failure: %v", err)
	}

	// Repositories
	userRepo := repository.NewUserRepository(db)
	iRepo := repository.NewInteractionRepository(db)
	metaRepo := repository.NewMetadataRepository(db)
	leadRepo := repository.NewLeadRepository(db, iRepo, metaRepo)
	apiKeyRepo := repository.NewApiKeyRepository(db)

	// Handlers
	authHandler := handlers.NewAuthHandler(userRepo, cfg.JWTSecretKey)
	leadHandler := handlers.NewLeadHandler(leadRepo)
	interactionHandler := handlers.NewInteractionHandler(iRepo)
	metaHandler := handlers.NewMetadataHandler(metaRepo)
	apiKeyHandler := handlers.NewApiKeyHandler(apiKeyRepo)
	webhookHandler := handlers.NewWebhookHandler(apiKeyRepo, leadRepo)

	router := gin.Default()
	router.Use(middleware.CORSMiddleware())

	// Health and Root
	router.GET("/api/health", func(c *gin.Context) {
		c.JSON(http.StatusOK, gin.H{
			"status":  "ok",
			"service": "BizCopilot Universal CRM",
			"version": "2.0.0",
		})
	})

	router.GET("/", func(c *gin.Context) {
		c.JSON(http.StatusOK, gin.H{
			"service":      "BizCopilot Universal CRM API",
			"docs_url":     "/docs",
			"swagger_url":  "/swagger/index.html",
			"health_check": "/api/health",
		})
	})

	// Interactive Swagger UI at /docs and /swagger/*any
	swaggerHandler := ginSwagger.WrapHandler(swaggerFiles.Handler)
	router.GET("/docs/*any", swaggerHandler)
	router.GET("/docs", func(c *gin.Context) {
		c.Redirect(http.StatusMovedPermanently, "/docs/index.html")
	})
	router.GET("/swagger/*any", swaggerHandler)
	router.GET("/swagger", func(c *gin.Context) {
		c.Redirect(http.StatusMovedPermanently, "/swagger/index.html")
	})

	// Public Routes
	api := router.Group("/api")
	{
		api.POST("/register", authHandler.Register)
		api.POST("/login", authHandler.Login)
		api.POST("/auth/register", authHandler.Register)
		api.POST("/auth/login", authHandler.Login)
		api.POST("/webhooks/lead", webhookHandler.IngestLead)
	}

	// Protected Routes (JWT Required)
	authMiddleware := middleware.AuthMiddleware(db, cfg.JWTSecretKey)
	protected := router.Group("/api")
	protected.Use(authMiddleware)
	{
		// Auth user profile & team
		protected.GET("/me", authHandler.GetMe)
		protected.GET("/auth/me", authHandler.GetMe)
		protected.GET("/users", authHandler.ListUsers)
		protected.GET("/auth/users", authHandler.ListUsers)

		// Leads static / collection routes
		protected.POST("/leads", leadHandler.Create)
		protected.GET("/leads", leadHandler.List)
		protected.GET("/leads/export", leadHandler.Export)
		protected.POST("/leads/bulk", leadHandler.BulkCreate)
		protected.PATCH("/leads/bulk", leadHandler.BulkAction)
		protected.POST("/leads/import", leadHandler.Import)

		// Custom Fields Metadata
		protected.GET("/leads/custom-fields", metaHandler.ListCustomFields)
		protected.POST("/leads/custom-fields", metaHandler.CreateCustomField)
		protected.PATCH("/leads/custom-fields/:field_id", metaHandler.UpdateCustomField)
		protected.DELETE("/leads/custom-fields/:field_id", metaHandler.DeleteCustomField)

		// Statuses Metadata
		protected.GET("/leads/statuses", metaHandler.ListStatuses)
		protected.POST("/leads/statuses", metaHandler.CreateStatus)
		protected.DELETE("/leads/statuses/:status_id", metaHandler.DeleteStatus)

		// Sources Metadata
		protected.GET("/leads/sources", metaHandler.ListSources)
		protected.POST("/leads/sources", metaHandler.CreateSource)
		protected.DELETE("/leads/sources/:source_id", metaHandler.DeleteSource)

		// Tags Metadata
		protected.GET("/leads/tags", metaHandler.ListTags)
		protected.POST("/leads/tags", metaHandler.CreateTag)
		protected.DELETE("/leads/tags/:tag_id", metaHandler.DeleteTag)

		// Duplicate Settings
		protected.GET("/leads/settings/duplicates", metaHandler.GetDuplicateSettings)
		protected.PUT("/leads/settings/duplicates", metaHandler.UpdateDuplicateSettings)

		// Saved Views
		protected.GET("/leads/saved-views", metaHandler.ListSavedViews)
		protected.POST("/leads/saved-views", metaHandler.CreateSavedView)
		protected.DELETE("/leads/saved-views/:view_id", metaHandler.DeleteSavedView)

		// Individual Lead Routes
		protected.GET("/leads/:number", leadHandler.Get)
		protected.PATCH("/leads/:number", leadHandler.Update)
		protected.DELETE("/leads/:number", leadHandler.Delete)
		protected.POST("/leads/:number/restore", leadHandler.Restore)
		protected.POST("/leads/:number/tags", leadHandler.AddTags)

		// Lead Notes
		protected.GET("/leads/:number/notes", interactionHandler.ListNotes)
		protected.POST("/leads/:number/notes", interactionHandler.CreateNote)
		protected.DELETE("/leads/:number/notes/:note_id", interactionHandler.DeleteNote)

		// Follow-ups
		protected.GET("/leads/:number/follow-ups", interactionHandler.ListFollowUps)
		protected.POST("/leads/:number/follow-ups", interactionHandler.CreateFollowUp)
		protected.PATCH("/leads/:number/follow-ups/:follow_up_id", interactionHandler.UpdateFollowUp)

		// Activities Timeline
		protected.GET("/leads/:number/activities", interactionHandler.ListActivities)

		// API Keys Management
		protected.POST("/api-keys", apiKeyHandler.Create)
		protected.GET("/api-keys", apiKeyHandler.List)
		protected.GET("/api-keys/:key_id", apiKeyHandler.Get)
		protected.POST("/api-keys/:key_id/revoke", apiKeyHandler.Revoke)
		protected.DELETE("/api-keys/:key_id", apiKeyHandler.Delete)
	}

	addr := fmt.Sprintf(":%s", cfg.Port)
	log.Printf("BizCopilot Universal CRM (Go+Gin) listening on http://localhost:%s", cfg.Port)
	if err := router.Run(addr); err != nil {
		log.Fatalf("Server startup failed: %v", err)
	}
}